Remove `osrs calc farm herb` subcommand
This is a really complex command that provides a ton of functionality but is annoying to configure and get everything right. It also adds a *ton* of code. So instead I turned it into a RuneLite plugin: https://runelite.net/plugin-hub/show/herb-farm-calculator

We don't need this code here anymore now, and it prevents us from running the CLI in the browser so I'm killing it.